Move 'Training' section above content in Timeline blocks
This increases the size of block titles so that it's larger than the Training header, and switches the position of the assigned trainings. It's a change that will aligns more with the intention for the trainings, which are preparation for the assignments described in the blocks. It'll also make way for our new 'Exercise' links for the Wiki Education wizard timelines for Fall 2018, which move much of the default timeline content from the blocks themselves into 'Exercise' training modules that are linked from the block content.