Use markdown_github option for pandoc conversions
GitHub-style markdown handles some syntax more in line with user expectations (and closer to MediaWiki markup) than the pandoc default. In particular, bullet lists without a preceding space are a common pattern, but Pandoc default does not interpret that as a list (unlike GitHub and the markdown-it JS library used by the Dashboard frontend). This should make on-wiki rendering of course descriptions more consistent.

Resolves #1946