Rewrite IndividualStatisticsPresenter calculation procedures
This fixes problems with double-counting of word count totals, and also cuts out a lot of extraneous queries for the earliest revision of an article for view sum purposes.

With all the double-counting issues taken care of, it should make it practical for us to include tracking-focusing courses on the dashboard for cases where we want to include the same user in two overlapping programs.