Move upload render logic out of componentWillReceiveProps
This sorts out some inconsistencies between redux state and rendering. Since componentWillReceiveProps is only called when props change, and not when a component is first mounted, the uploads were behaving inconsistently when you visit the Uploads tab, then another tab (thereby unmounting the UploadsHandler), and then return. This was masked by reloading the uploads each time the UploadsHandler re-mounted, since it would re-receive the uploads data (and then need to refetch the metadata as well).