Add debug logging for UpdateCourseStats
On P&E Dashboard, the WikiProject Wikidata India program has not updated in 20 days. There are no failed jobs, so my guess is that the process is running out of memory and dying without rescue because there are so many records. This should help pin down where the problem is.