Switch from Trix to TinyMCE for rich text editing
Trix is pretty limiting in what it can do, and in particular, it does not retain CSS classes for the HTML elements if any change is saved, which means it strips out the special formatting for exercise links in the timeline. TinyMCE is a mature alternative, and the react integration seems to work smoothly.