Add animation for when course stats change
This adds logic for detecting when updated course stats come back, and adding a keyframe animation to the numbers when that happens. It also fixes the stats update polling, which was broken with the recent switch to Redux for course state.

It also drops the polling interval to 2 minutes, so that two polls in a row with updated data won't cause animations to be skipped.

This should make it more interesting to leave a course page open during an editathon.

Resolves #1853