Improve the return behavior at the end of training modules
If a user reaches a training module from a course page (or anywhere else within the app), they should be returned where they started from. This saves the referrer in the user's session and passes it to React so that the 'Done' link can point back to the course page when possible.