Fix mass addition of users for non-normalized usernames
MassEnrollmentWorker was hitting errors for users that are added with a non-normalized form of the username, because the original input was used as the key in the results hash rather than the normalized version of the username that we store in the User record and find the user by.

This refactor of AddUsers puts the normalized username in place of the submitted text, so that the usernames of successfully-added users alway match the User records.