If users log in via enroll URL, redirect straight to enrollment
When a user visits a course URL that includes the 'enroll' parameter, this generally means they followed an enrollment link distributed by the instructor or event organizer. If they don't have an account or aren't logged in, they will first need to sign in via OAuth before they can actually join the course. They would be redirected to the original enroll URL upon login, but they still needed to click 'Join' to actually finalize their enrollment. Users who didn't do so would end up logged in on the Dashboard but not enrolled in their course, a relatively common and confusing situation.

With this change, we use data stored in the user's session to handle cases where a user has logged in starting from an enroll link, and sending them straight to the enroll URL, bypassing the need to click 'Join'.