Fix unnecessary re-rendering of StudentList
The filtering of assignments in the students loop of StudentList instead of within each Student component meant that every Student component got a new array (with the same contents) every time StudentList rendered. Moving that filtering step to the Student component means that the Students will only re-render if their props actually change.