Avoid setting passcode to the '****' placeholder
If a user saves a course that received the placeholder passcode ('****'), this should be treated as not having a passcode in the saved data at all. The placeholder is received if the user does not have a current session, which may happen if they are switching between browsers and the course data (including passcode) automatically refreshes in an open window that subsequently gets saved.