Clean up training loose ends
Setting the initial completed state to null instead of false is used to avoid briefly flashing the 'wait' banner when the final slide is loaded directly.

The MODULE_COMPLETED action doesn't need to be separate from SLIDE_COMPLETED, as the server response always includes whether or not the module has been completed.

There were a few stray semicolons being rendered as text.

The isFirstSlide bit of state was completely unused.