Quick fix for broken survey flow when conditional questions change
This works around a situation where changing the answer to a question upon which the next question depends would break the survey. The included spec verifies that it's possible to complete the survey after returning to a condition question and changing the answer, but it depends on the new behavior of just always leaving the conditional question in place after the condition changes. It would need to be changed if resetConditionalGroupChildren is fixed to remove the conditional question without breaking the survey flow.

Resolves #1263