In training index, emphasize the most relevant library
This adds variant styling for a logged in user, based on their most recent course, so that the training index emphasizes the set of training modules that are relevant for their course. This will hopefully cut down on cases of users taking the wrong version of an assigned training module — students vs professional development.