Match errorQueue behavior of ValidationStore in reducer
ValidationStore.isValid() both checks for invalid entries and inserts the error messages for any validations that were initialized as invalid but never updated. This behavior was exploited in CourseCreator so that the first time you click 'Create my Course!', all the blank fields are highlighted as invalid and an error message is displayed.

In Redux, that needs to be an explicit action that changes the state: ACTIVATE_VALIDATIONS.