Redirect users to home after a manual locale update
The language picker is not shown on Wiki Education Dashboard, but some users with browsers set to a different language need to be able to set their locale to English (or something else), and that is done via /update_locale/en. The resulting json success message is confusing; redirecting to the home page matches the user experience for changing the locale via the language picker (which refreshes the page after success).