Improve method of switching between local and git for engines
The environment variables were more trouble than they were worth. We can just comment/uncomment the one we want, which plays nicely with guard also because after you are done working on the engine and push changes there, then when you switch back to the github source it will update the Gemfile to the newest commit upon `bundle install`.