Ticketing system
Add ticketing system for admins which included the following commits:

* Add Ticketing & Messages migrations

* Add basic testing and validations for Ticket & Message models

* Update ticket and message schema

* Add faker gem

* Add rake task to populate demo ticket data

* Add ticket controller

* Add basic frontend for listing tickets

* Swap TicketDispenser engine in for ticketing features

* Protect tickets routes

* Minor updates

* Sort tickets by table headings

* Add Griddler gem

* Allow for incoming email

- Current WIP for allowing incoming email
- Requires the use of a Wikipedia Expert special user(s)

* Set ticket sender and owner

* Refactor and test email_processor

* Add show page for tickets

* Reply to message

* Send email notification on new replies

* Fix navigation from index to show ticket, clear form on submit

Render runs before componentDidMount, and that's what sets the selected ticket. So we need to bail before rendering the ticket until the selelected ticket has been set.

* Update use of TicketDispenser

- Use Ticket Dispenser service
- Add ticket on NeedsHelpAlert
- Other minor modifications

* Add tickets, account requests to admin navbar

* Pull NotificationsBell to a separate component

* Ring NotificationBell for open tickets as well as requested accounts

Now the bell goes to /admin, but we want to still go to the /requested_accounts page directly for Programs & Events Dashboard, where tickets aren't yet going to be used.

* For P&E Dashboard, do not surface ticket notifications

* Minor style updates to email

* Change name of TicketDispenser folder

* Read messages when viewing a ticket

* Remove loading key for tickets reducer

* Fade ticket row when all messages read

* Update admin object in html

* Pull TicketDispenser from GitHub

* Add note to status definitions for tickets

* Update email processor tests to rely on greeter instead of special users

* Update code to use project instead of course

- TicketDispenser now uses a generic project label instead of
  specifying course

* Update to latest ticket_dispenser

* One more course_id -> project_id

* One more spec fix

* Do not set a starting value for ticket table sorting

* Remove ticketing rake task

* Don't require task with Faker dependency

* Fixing loading indicator for when there are no tickets

* Update ticket status when sending reply

This works around a couple of problems. With two different form buttons, we don't go through the form `onSubmit` anymore, so we can't use the clearOnSubmit feature of TextAreaInput. Instead, it now uses the latest message as a key, to generate a fresh input after a message is added successfully.

It also adds a guard against submitting an empty reply. This only works with the plain text version of the editor, rather than the HTML stored as `content`, since it will turn whitespace into `<p>`, `nbsp;`, etc. To get the plain text, we pass the editor onChange event back through InputHOC as a third argument to the provided onChange function... which we can then use to access the editor itself and get the plain text version of the content.

* Fix Gemfile.lock to point to GitHub

Whoops, bundler undid the change I thought I was committing, since it didn't have the same ENV variable that I set in my terminal.

* Make the 'Resolve' button work for tickets

Also, move the CSRF token workaround into ticket_actions.js.

* Remove unused constant in tickets reducer

* Do not mutate state in tickets reducer

* Change Resolve button to Select

- Handle ticket status via a select box

* Improve method of switching between local and git for engines

The environment variables were more trouble than they were worth. We can just comment/uncomment the one we want, which plays nicely with guard also because after you are done working on the engine and push changes there, then when you switch back to the github source it will update the Gemfile to the newest commit upon `bundle install`.

* Scope query of open tickets to current user or unowned

Also, show owner by username rather than real name.

* Refactor SpecialUsers and Setting, fix wikipedia_expert downgrade

This moves the setting of special users out of Setting and over to the SpecialUsers class instead, as Setting doesn't need to have any knowledge of the SpecialUsers implementation build atop it.

The if/else branches for handling `wikipedia_experts` as a group role is a little messy, but I can't think of a better solution at the moment.

* Add selector to change Ticket owner

The set of all admins is passed through the react_root HTML and inserted into redux, so that any admin can be chosen as the new owner of a ticket.

* Fix specs after refactor of Setting and SpecialUsers

* Handle unowned tickets

* Move ticket status and owner selectors next to related data

There's room to refine the layout further and reduce duplicate information, but this will work for now.

* Appropriately thread emails from responses