Fix order of asset upload and symlinking during deployment
This puts the switching of `current` to the ongoing deployment immediate after the assets are uploaded, so that the app will continue serving the old assets until the new ones have been uploaded and the bandwidth-constrained steps are complete.

Before this change, slow asset uploads would cause a period where assets are not available and the website is basically broken.