Add stats download modal to Campaign overview
This adds the CSV downloads for a campaign, previously only available on the campaigns index, to the overview page of each campaign. The new feature is implemented with React, so it adds the start of a CampaignOverviewHandler component for gradually porting the rest of the view to React.

Also, it's implemented with React Hooks, the first use of Hooks in the codebase!