Don't post to talk page if assigned article is deleted
This is a quick fix for a bug in which a student may end up assigned to an Article record marked deleted. If there are any assignments to the non-deleted version of the same Article, then they would be treated as two separate articles and so each update would overwrite the last one.