Add missing enrollment templates during 'greet' routine
This performs the same set of edits normally attempted by an enrolled students account right after enrollment, whenever a Wiki Education staff member uses the 'greet ungreeted students' routine to send a welcome message to each student in a course. The templates only get added if they are not already present, so this serves as a fallback for when the initial enrollment edits failed, which could happen because of CAPTCHAs for a new account.

This will only add the templates for 'ungreeted' students, meaning students whose talk pages have not yet been edited by one of the greeters, so at worst if some edits are failing repeatedly on the same page, this will only attempt to add each template one additional time.