Adapt CourseStats for Wikidata and other non-Wikipedia courses
This is the start of making CourseStats more flexible based on the course stats present and the home wiki. It hides 'bytes added' for Wikidata courses, and updates the labels for Wikidata and for other non-Wikipedia courses, and also hides the uploads count unless there is at least one upload.