Add redirects to English Wikipedia sandbox pages
This lets us include static links in training modules that will resolve to a user sandbox page on Wikipedia, without relying on Special:MyPage. Special:MyPage links often take students to the wrong place because they are logged in on the Dashboard but not on Wikipedia. With this, they'll end up at the right sandbox page on Wikipedia even if they are logged out.

Query params (like veaction=edit&preload=Template:Dashboard.wikiedu.org_choose_article) will be included as well.

This can be adjusted later to support other wikis, if necessary.