Ensure that WikiApi#get_page_content always returns a UTF8 string
We started seeing errors from the routine that posts enrollment templates to userpages. In particular, it seems to break when a userpage includes some non-ascii characters and the course slug within the template does as well. This causes an encompatible encoding error when the code checks for whether the template is already included on the page.

This was happening because WikiApi#get_page_content returns an ASCII-8BIT string. Normally Ruby handles this without a problem; as long as either string contains only ASCII characters, string_one.include?(string_two) works as expected. But if both contain non-ascii characters, it blows up. The solution is to make sure that we convert the strings to UTF8. Fortnately, they are actually UTF8 strings that we get from MediaWiki API, so this change produces the correct content rather than the `/x**` encodings showing up in the strings.