First pass as switching to 'Words Added' for Students and Articles
We convert characters/bytes to words to give course-level estimates of content added, based on a value estimate for English Wikipedia. We want to do this also at the level of individual articles and editors as well, but without disrupting stats for other wikis where we don't have an estimate of bytes per word.

This patch addresses the issue by adding a `bytes_per_word` method for Wiki, which will return nil for any wiki that we don't have an estimate for (ie, everything but en.wikipedia.org). The strategy here is to pass that value for the home wiki of the course to the frontend, and then switch between words and bytes based whether a `home_wiki_bytes_per_word` value is received. This way, it should be easy to extend support for other wikis if/when we have bytes-per-word estimates for them.