Update TrainingSlide table to support 4-byte unicode characters
This adds 4-byte unicode support to the 'content' and 'translations' fields of TrainingSlide, and fixes the 'prohibition sign / no symbol' on several slides.

See https://en.wikipedia.org/wiki/No_symbol

We've been using the "Combining Enclosing Circle Backslash" for this, but it doesn't render in Chrome (and requires extra spaces as a workaround, since it is supposed to render on top of the previous character). The "prohibition sign" is the same symbol, but it's a 4-byte character that requires schema changes to work.

This change will also require changes to the database.yml `encoding` setting; it should be changed in production before deployment (and will need to be changed in local development environments as well).