Change final button from 'Done!' to 'Okay' in exercise modules
For exercise-type modules, the 'Done!' button is a source of confusion for students, implying that clicking 'Done!' at the end of the module will update the exercise status to 'Complete'. Instead, clicking 'Mark Complete' from the Home tab is necessary.

The modules should also be updated to explicitly instruct students to use the 'Mark Complete' button after the task is done, but this will be a good start.