Use default behavior of wiki edits toggle for FellowsCohort
This removes the hard-coded 'true' for wiki_edits_enabled so that setting wiki edits to false on the front-end has the expected behavior. Otherwise, you can set it to 'no' and save, but upon refreshing the course page it will show up as 'yes' again, and (unless the enrollment edits toggle was changed also) enrollment edits will still happen.

With this change, there's still some mismatch between the edit settings UI and the behavior, because you can set the course page edits or assignment edits to 'yes' and save, but they will still go back to 'no' upon reloading and cannot actually be enabled.