Cut out unused drawer-toggling code from StudentList
The StudentList no longer needs to include drawers at all, so we can cut out all the props and code that deal with drawer-toggling.

Also, this refactors ContentAdded to just render the contents of a cell rather than an entire <td>, so that the onClick behavior for that cell works the same as the rest of the cells (aside from the uploads one) in a StudentRow.