Drop the Replica concurrency limit to 2
It looks like a current bottleneck in our system is the ArticleStatusManager step of daily updates; the sidekiq process dies soon after starting, possibly because the system runs out of memory with 10 threads at once. This is a shot in the dark to see if too many threads is the thing causing the system to become unstable.