Add choice of course format (in-person vs online) to CourseCreator
This adds a new 'format' flag to ClassroomProgramCourse-type courses. It also includes a tweak to the way the course slug is displayed to that the Next button doesn't overlap with the now-bigger right column of the creator form, and it improves the course creation spec to check for the values of the course level, format, and subject. (It also removes some dead code in that spec.)