Added error handling module.
This module will take an error and some additional parameters and accordingly log the errors to sentry. Currently, UpdateServiceErrorHelper uses it to pass sentry tags to keep track of course errors. It is named ApiErrorHandling because this module may be expanded to error handling and all sorts of those things in the future. For now, it logs errors and checks error levels by dynamically getting the TYPICAL_ERRORS that occur in the APIs.