Error Helper for update service UpdateCourseStats.
This module will help keep track of sentry_tag_uuid, the error_count and getting sentry tags for tagging data that is passed on to Sentry. update_service_id distinguishes each update and course slug is for the course. So basically now we can filter errors specific to course and a specific update process also.

The reason why we chose a pass down bottom strategy for tracking course update errors was that it was seemingly better than the other alternative pass up strategy. In the pass up strategy, we would have to put rescue blocks all over the codebase whereever any method or whatever in the upward path was being used anywhere else than the course update process. It also led to a more brittle, easily breakable course update even on one error. While on the other hand passing the course down helped us in the way that arguments could be optional and we had a work around for keeping logic in one place by extracting error handling code to new modules.

Also, firstly we tried a technique of saving CourseErrorRecords in the database for them to be shown earlier, but now that we have a method of uniquely identifying errors in Sentry UI and that the ultimate goal is to make the errors more discoverable by end users.

Now we don't have any database constraints and easily keep track of errors.

In summary we tag each course update error using a unique update process uuid and course slug and then send it to sentry in ApiErrorHandling module.