Using UpdateLogger to save update process info.
We will use UpdateLogger.update_course to store the update process summary regarding the error counts and sentry tag which stores this information in the course flags for the lst 10 updates.

One thing is to be kept in mind that sentry_tag_uuid and error_count resets whenever UpdateCourseStats.new is called. They are a part of the module UpdateServiceErrorHelper.