Added spec for currently running jobs, tests find_active_job path.
Refer to Sidekiq API Workers for information about the hash value returned, as is mocked by the tests. There were two options, either to run the background job by inserting rake commands programatically or mock the values. We just want to ensure the correctness of our function, for which mocking is enough.