Reflected new changes involving fetching course specific revisions.
The isCourseScoped (earlier isCourseSpecific) is not really something that should go into redux state. The only thing that cares about it is the component and its toggling part so that it can make the requests accordingly.
Also, introduced the new variable redux state attributes as done in the new changes.