Passing course future update criteria to frontend throuugh json.
The ready_for_update scope queries the whole database for the courses which are to be updated, but we want to know the same for only one course without that inefficient overhead. Made a new method to pass on whether or not the course is to be updated in future through course.json.