Displaying loading on each request, preventing overlapping requests
We do not display the shoe more button if there is a request which is still resolving, thereby preventing another request to be flagged off until the initial one is resolved. Also, we display loading on each request hit and do not display the revisions_none message until we are sure that there are no revisions (i.e., when the revisions are loaded).