Removed condition for newly created articles.
Newly created articles are updated up to state on creation unlike categories which has empty list of articles on creation, so for articles, they are already updated on creation and do not require to be special cased as in the updation of categories. Also added updated_at: 2.days.ago in specs for update_article_status_for_course reflecting the new behaviour that from now onwards, only articles updated more than 1 day ago will be updated.