Made another method for getting titles without prefixes.
This is because page_titles_for_category is used at many places which might be requiring the prefixes.
Also, changed method arguments to make property parameter optional. This is done because in case of removing the prefixes from titles we also want data regarding namespace for each page ('ns'), so in case of mainspace_page_titles_for_category we don't pass anything in property parameter and hence get all the data of the page received from API, which we then pass on to CategoryUtils to appropriately format.