Returning all page data instead of page data properties.
Earlier, only title property from the page data received from API was returned, then I added two options to return full page or a specific property as and when required. This is confusing and irrelevant and should not really be the task of get_category_member_data. Moved extracting of titles to CategoryUtils with appropriate documentation of API response.