Add support for parsing a wiki page representing a TrainingLibrary
Once fully implemented, this should let us move away from .json pages for defining the training libraries, and to enable translation of the full gamut of TrainingLibrary content.

This commit just extends WikiSlideParser to handle the new template I created for holding the 'categories' data of a TrainingLibrary; the loader needs to be updated to use it now.