Get rid of inaccessible and deprecated chosen-js
This removes the JavaScript UI on multi-select inputs in the Survey admin interface, such as the pickers for the campaigns and tags for a Question Group. This means the multi-select inputs will now be unstyled, and selection of multiple options requires using CTRL+click.

Ideally, we'd replace this with a well-maintained and accessible way to style these inputs and provide a nice UI. But it's not crucial, as long as they are functional and accessible via screen reader.