Avoid reflow when scroll buttons render
When the arrow icon isn't already cached in the browser, the alt text for the scroll button would flash briefly before being replaced by the arrow, and cause a significant jump in position. Setting the height/width prevents reflow, and setting 'color: transparent' on the img prevents the first few letters of alt text showing up before the image loads.