Update Faraday and Sentry
The mediawiki_api gem specifies a pre-1.0 version of Faraday and related gems, which prevented us from using the newest version of Faraday and Sentry. I've forked it to remove the limitation to verify that the gem works as expected with Faraday 1.1 before I try to merge the change upstream.