Refactor and rename Course#new_articles
This method actually returns ArticlesCourses, which renaming makes clear.

Also, the join to `articles` to ensure namespace 0 is not necessary, because (unlike when this was first implemented) ArticlesCourses are exclusively used for mainspace articles (and get destroyed if the article ends up moved out of mainspace). The situations where we retain ArticlesCourses records for articles that aren't counted toward 'articles edited' totals are for untracked articles and for deleted articles (I think).