Add Book namespace
Even though this is marked as unused, it appears to be technically possible for a page to exist in the Book namespace (108). We have an Article record in the database that has this namespace (which has since been moved to mainspace), so this is necessary to prevent problems when generating data about that record.