Skip article status updates for resource-intense courses
This will hopefully cut down on the load caused by very active courses on Programs & Events Dashboard without materially affecting the results. For courses that edit thousands of articles, the impact of deletions and namespace moves is likely to be fairly limited, and this the trade-off in catching those things versus speeding up updates and avoiding situations that break the app altogether is worth it.