Use longest update time instead of average for queue sorting
After the recent changes to the update process — where courses in the 'long' queue skip the time-consuming step of checking for deletions and page moves among all the edited articles — we now have very few courses in the 'long' queue, and the 'medium' queue latency is up to 5 days. This is likely because slow-to-update courses move into the medium queue after spending a short amount of time in the long queue, and don't move back to the long queue immediately after one longer update. This slight tweak will hopefully shift the balance closer to having a 'long' queue with high latency and a 'medium' queue with at most a day of latency.