Change behavior for transient OAuth errors
We've been running into an issue lately where admins get logged out when attempting to greet the students in a course, because the Dashboard hits a 'Nonce already used' error when attempting to make the greeting edits.

Since that indicates a collision of edits rather than completely invalid credentials, we shouldn't set the token invalid (and thereby log the user out of the Dashboard) when this happens. Other variations of 'mwoauth-invalid-authorization' are still problems that require us to log the user out and in again.