Prevent edits on wikis where edits are not explicitly enabled
Turns out the student greeting edits could still happen on another wiki besides en.wikipedia if that course's home wiki is not en.wikipedia.

This adds a proper guard for the 'GreetUngreetedStudents' service, and also adds a final check to surface anywhere else we might have missed that could attempt to send edits to a wiki that isn't explicitly configured to allow edits.