Workaround for JSCover coverage collection in wizard
When instrumented to collected test coverage data with JSCover, window.onbeforeunload is used to persist coverage data to local storage. This is why we could never capture coverage data for the wizard: submitting the wizard would set it to `null`. As a workaround, we cache any onbeforeunload that was already there and then restore it, instead of just setting it to null when the wizard has been submitted.