Improve separation of slow-updating courses from typical ones
This will push the system back towards prioritizing not-too-slow courses. Now the 'long_update' queue will consist of every course that has ever taken more than 10 minutes for an update — tracked by the :longest_update flag — and the limiting of article status updates will be done based on the longest recent update, so that the medium_update queue isn't clogged with the article status updates for slower courses.