Start moving CSV generation to a background job
This changes the behavior of campaign CSV generation — just the 'courses' CDV so far — so that it happens in a background job that doesn't tie up Passenger while the data is being generated. It will return a 'being generated' message initially, and then give the file once it is done being generated.