Improve username collision handling in UserImporter
When we have two records for the same user (because they changed their username, but the new username was added to the Dashboard before the old one got updated, there's a duplicate key collision when we try to fetch the global_id and registration date for the old username. This ensures that the old username gets removed from the system so that we don't have to perpetually retry.

There are two such cases on Programs & Events Dashboard right now, which this should take care of.