Use batches for updating average views
For courses with tens of thousands of articles, this was a lot of data to collect before persisting any of it. Reasonably sized batches should get most of the efficiency of the thread pool approach, but working with fewer records at a time and writing to the database more frequently.