Only use 1 OAuth ID for replica queries
I've changed application.yml on Wiki Education Production to only use the current OAuth ID as well. Using more than one results in a join that duplicates revision results (one for each change tag) and is likely the cause of inconsistency in the accuracy of the `system` attribute on Wiki Education production.

This was surfaced via a failing replica spec because the order of results changed upon MariaDb version bump when the Replica databases switched over to new versions. See https://phabricator.wikimedia.org/T278983