Reimplement SurveysQuestionGroup.by_position to return a relation
This is a functional change in how the .by_position method works, as it would previously return (first) all the SurveysQuestionGroup records for a given survey ID ordered by `position`, and then follow it with rest of the records that have no `position` value ordered by `created_at`. I think this is a SurveysQuestionGroup doesn't have a position unless the set of them for a survey gets reordered; this ensured that newly added ones would go at the end, while preserving the order of the ones that were already there.

As such, this means we have a couple of early surveys in the Wiki Education Dashboard database where the order will be different now. However, we've basically moved to a system of only using one QuestionGroup per survey, so this order change shouldn't affect any recent surveys.

Having a relation means we can also use `includes` when rendering Survey results. This is version is just a guess at what records are being used in the results view, and probably isn't a complete solution to the N+1 queries afflicting the results view.