Update to Omniauth 2
Omniauth 2 requires that all signin requests use POST and include an authenticity token. The easiest way to implement this is to use gem "omniauth-rails_csrf_protection" and use `data-method="post"` in combination with the rails-ujs JavaScript helper to convert anchor links into POST requests.

This change also means we can't redirect to a login URL, which we were doing as part of the self-enrollment flow for a logged out user who visits the "enroll_self" route directly. However, I don't think there's an actual use case for this, as the distributed enrollment links don't redirect a user to this route until they are already logged in. I'm not aware of any failure cases in the enrollment flow where this is actually needed; it might just be a relic of the older enrollment flow before we implemented the EnrollCard UI. As such, we can just return a 401 if the enroll_self is visited while logged out.