Disable drag-and-drop test
I tried the workarounds from the issue thread, but still couldn't find a way to get this test to work with a current version of react-dnd. https://github.com/react-dnd/react-dnd/issues/1195

However, the drag-and-drop feature itself works fine (with slightly nicer UI, as the dragged element shows up both in the mouse-hand and in ghost form in the position it would take upon dropping it.