Fix some warnings
The form builders for the campaign programs list take the campaign, so the default IDs for each input are not unique for each course row. This results in elements with duplicate IDs.

Also, now the `raise_on_missing_translations` setting is supposed to be done via `config.i18n` so that it can catch missing translations at both the controller and view layers.