Add module status or "Mark Complete" button to Timeline and ThisWeek
With this change, everywhere that links to a training module or exercise will include the status of the current user, if they are a student. In particular, this ensures that the "Mark Complete" button shows up on the Timeline to prevent confusion in cases where that's the place a student launched the exercise from.