Specify entrypoint as "launcher" for Nomad job tasks
This makes the jobspec compatible with Waypoint 0.3.2 and up. Waypoint 0.3.2 included a version bump of `pack`, which in turn bumped `lifecycle`, which introduced setting the `web` process as the default process type for the image.

Adding a default process type meant that Nomad would use the entrypoint for the default process type unless another entrypoint is specified. So instead of simply running the command for the sidekiq task, it would run the default web entrypoint and use the `command` as arguments.

We intend for the `command` and `args` to completely specify how to launch the container for any each task, so we specify the base buildpack `launcher` command as the entrypoint for each job.

Relevant context: https://github.com/hashicorp/waypoint/issues/1579