Adjust the timing of UnsubmittedCourseAlert emails
Now that we have a submission deadline before the start of the term, we want to send these out significantly earlier; we send them out if a course is unsubmitted after two weeks.

This also includes a cutoff to avoid sending alerts for the many older courses that were never submitted for previous terms.