Fix conditional question hiding in surveys
It looks like there was a bug with the data-conditional-question attribute introduced with the conversion of the _question_group template from ERB to HAML, which broke the behavior of conditional questions that aren't supposed to show up for users unless the required conditions are met from previous answers. The attribute was instead being interpolated into the `class` of the question group div. This slipped through because we didn't have any tests that covered the situation of a conditional question that shouldn't be rendered.

The included extension of the surveys spec covers this feature (and the test fails without the fix).