Rename "stage" to "subject" in course advice email code
This refactor paves the way for extending CourseAdviceMailer to handle additional advice emails that aren't tied to a specific stage of the course. In particular, it will be used for biography advice emails that will be sent at the time of course approval, and we'll also likely fold course approval and course approval followup emails into it later (to share code for who receives such course-related emails).