Rewire how courseCreationNotice gets passed around
With this change, we load the Deadlines setting record for each call to .course_creation_notice, so that it's always current even with multiple application servers. To make sure we don't do that query more than needed, this change puts the message into the react root of the two views that need it (course#show and course_creator).