Disable Growth experiment features for newly-created accounts
This will make sure students who create accounts via a signup link on the Dashboard don't end up with experimental features that may cause confusion due to a different user experience from classmates and/or Wiki Education training material.

Once the Growth features (such as the newcomer homepage) are default for English Wikipedia, we can undo this and rely on the default new user settings (and update training materials accordingly, if necessary).