Fix buggy handling of sparse follow-up answers
The code for rendering follow up answers in survey results assumed that the answers and the follow-up answers were arrays of the same length. In some cases that isn't true (for example, because only some answers include follow-up answers, and/or some responses include a follow-up answer but not an answer to the main optional question). In those cases, the survey results page could display the wrong username for a given follow-up response.

This is a quick, minimally tested attempt to sort out the behavior by providing the follow-ups as an object with answer IDs as the keys, instead of an array.